Initial commitment for Zarex - name change and comments (not fully).

+++ Zarex Server/src/com/netty/model/combat/MagicCombat.java
@@ -3,9 +3,6 @@ package com.netty.model.combat;
 import java.util.Random;
 
 import com.netty.model.Entity;
-import com.netty.model.combat.AttackType;
-import com.netty.model.combat.Combat;
-import com.netty.model.combat.CombatType;
 import com.netty.model.player.Player;
 import com.netty.model.player.skill.Skill;
 import com.netty.model.player.skill.SkillType;
@@ -14,12 +11,32 @@ import com.netty.model.update.Hit;
 import com.netty.model.update.HitType;
 import com.netty.model.update.UpdateFlag;
 
+/**
+ * 
+ * 
+ * @author Joshua Rodrigues
+ * @since Sep 21, 2011 10:33:33 AM
+ */
 public class MagicCombat extends Combat {
 
+	/**
+	 * 
+	 * @param entityPlayer
+	 * 			The id to set.
+	 * @param entityEnemy
+	 * 			The id to set.
+	 * @param combatType
+	 * 			The id to set.
+	 * @param attackType
+	 * 			The id to set.
+	 */
 	public MagicCombat(Player entityPlayer, Entity entityEnemy, CombatType combatType, AttackType attackType) {
 		super(entityPlayer, entityEnemy, combatType, attackType);
 	}
 
+	/* (non-Javadoc)
+	 * @see com.netty.model.combat.Combat#mageDamage()
+	 */
 	@Override
 	public void mageDamage() {
 		Random random = new Random();
@@ -49,17 +66,23 @@ public class MagicCombat extends Combat {
 		skill.addExperience(2 * damage);
 		this.getPlayer().setAnimation(new Animation(424, 0));
 		this.getEntityEnemy().setAnimation(new Animation(1659, 0));
-		this.getEntityEnemy().setHit(new Hit((byte) damage, HitType.NORMAL_DAMAGE));
+		this.getEntityEnemy().setHit(new Hit(damage, HitType.NORMAL_DAMAGE));
 		this.getEntityEnemy().getUpdateFlags().setUpdateFlag(UpdateFlag.HIT);
 		this.getEntityEnemy().getUpdateFlags().setUpdateFlag(UpdateFlag.APPEARANCE);
 		this.getPlayer().getUpdateFlags().setUpdateFlag(UpdateFlag.APPEARANCE);
 	}
 
+	/* (non-Javadoc)
+	 * @see com.netty.model.combat.Combat#meleeDamage()
+	 */
 	@Override
 	public void meleeDamage() {
 		return;
 	}
 
+	/* (non-Javadoc)
+	 * @see com.netty.model.combat.Combat#rangeDamage()
+	 */
 	@Override
 	public void rangeDamage() {
 		return;
