+++ Zarex Server/src/com/netty/model/combat/RangeCombat.java
@@ -11,22 +11,48 @@ import com.netty.model.update.Hit;
 import com.netty.model.update.HitType;
 import com.netty.model.update.UpdateFlag;
 
+/**
+ * 
+ * 
+ * @author Joshua Rodrigues
+ * @since Sep 21, 2011 10:32:59 AM
+ */
 public class RangeCombat extends Combat {
 
+	/**
+	 * 
+	 * @param entityPlayer
+	 * 			The id to set.
+	 * @param entityEnemy
+	 * 			The id to set.
+	 * @param combatType
+	 * 			The id to set.
+	 * @param attackType
+	 * 			The id to set.
+	 */
 	public RangeCombat(Player entityPlayer, Entity entityEnemy, CombatType combatType, AttackType attackType) {
 		super(entityPlayer, entityEnemy, combatType, attackType);
 	}
 
+	/* (non-Javadoc)
+	 * @see com.netty.model.combat.Combat#mageDamage()
+	 */
 	@Override
 	public void mageDamage() {
 		return;
 	}
 
+	/* (non-Javadoc)
+	 * @see com.netty.model.combat.Combat#meleeDamage()
+	 */
 	@Override
 	public void meleeDamage() {
 		return;
 	}
 
+	/* (non-Javadoc)
+	 * @see com.netty.model.combat.Combat#rangeDamage()
+	 */
 	@Override
 	public void rangeDamage() {
 		Random random = new Random();
@@ -56,7 +82,7 @@ public class RangeCombat extends Combat {
 		skill.addExperience(2 * damage);
 		this.getPlayer().setAnimation(new Animation(424, 0));
 		this.getEntityEnemy().setAnimation(new Animation(1659, 0));
-		this.getEntityEnemy().setHit(new Hit((byte) damage, HitType.NORMAL_DAMAGE));
+		this.getEntityEnemy().setHit(new Hit(damage, HitType.NORMAL_DAMAGE));
 		this.getEntityEnemy().getUpdateFlags().setUpdateFlag(UpdateFlag.HIT);
 		this.getEntityEnemy().getUpdateFlags().setUpdateFlag(UpdateFlag.APPEARANCE);
 		this.getPlayer().getUpdateFlags().setUpdateFlag(UpdateFlag.APPEARANCE);
